refactor awards editor in its own component
